requirements.txt
black==23.3.0
ipykernel==6.23.1
isort==5.12.0
kornia==0.6.12
matplotlib==3.7.1
numpy==1.24.1
opencv-python==4.7.0.72
pandas==2.0.1
Pillow==9.3.0
pycocotools==2.0.6
pytorch-triton==2.1.0+7d1a95b046
rich==13.3.5
scikit-learn==1.2.2
scipy==1.10.1
torch==2.1.0.dev20230516+cu118
torchaudio==2.1.0.dev20230516+cu118
torchmetrics==0.11.4
torchvision==0.16.0.dev20230516+cu118
tqdm==4.65.0
